impacts

+++ src/world/handler/impact.cpp
@@ -1,6 +1,6 @@
 #include "impact.h"
 
-Impact::Impact(WorldObject* worldObject, Voxel *voxel, const glm::ivec3 &vec):
+Impact::Impact(WorldObject* worldObject, Voxel *voxel, const glm::vec3 &vec):
     m_worldObject(worldObject),
     m_voxel(voxel),
     m_vec(vec)
@@ -15,6 +15,6 @@ Voxel *Impact::voxel() {
     return m_voxel;
 }
 
-const glm::ivec3 &Impact::vec() const {
+const glm::vec3 &Impact::vec() const {
     return m_vec;
 }
